+++ R/register-s3.R
@@ -41,8 +41,8 @@
 #' # printing when they are used.
 #'
 #' .onLoad <- function(...) {
-#'   s3_register("pillar::pillar_shaft", "vctrs_vctr")
-#'   s3_register("tibble::type_sum", "vctrs_vctr")
+#'   .s3Register("pillar::pillar_shaft", "vctrs_vctr")
+#'   .s3Register("tibble::type_sum", "vctrs_vctr")
 #' }
 #' @return nothing; called for side effects
 #' @keywords internal